refactor: move StackCheckPass to own file
