+++ riscv_analysis/src/lints/mod.rs
@@ -27,3 +27,6 @@ pub use callee_saved_register::*;
 
 mod lost_callee_saved_register;
 pub use lost_callee_saved_register::*;
+
+mod stack;
+pub use stack::*;
